Move events.publisher_statuses to publisher.statuses
