+++ rcljava/CMakeLists.txt
@@ -60,8 +60,8 @@ set(${PROJECT_NAME}_jni_sources
   "src/main/cpp/org_ros2_rcljava_contexts_ContextImpl.cpp"
   "src/main/cpp/org_ros2_rcljava_executors_BaseExecutor.cpp"
   "src/main/cpp/org_ros2_rcljava_events_EventHandlerImpl.cpp"
-  "src/main/cpp/org_ros2_rcljava_events_publisher_statuses_LivelinessLost.cpp"
-  "src/main/cpp/org_ros2_rcljava_events_publisher_statuses_OfferedQosIncompatible.cpp"
+  "src/main/cpp/org_ros2_rcljava_publisher_statuses_LivelinessLost.cpp"
+  "src/main/cpp/org_ros2_rcljava_publisher_statuses_OfferedQosIncompatible.cpp"
   "src/main/cpp/org_ros2_rcljava_node_NodeImpl.cpp"
   "src/main/cpp/org_ros2_rcljava_publisher_PublisherImpl.cpp"
   "src/main/cpp/org_ros2_rcljava_service_ServiceImpl.cpp"
@@ -133,8 +133,8 @@ set(${PROJECT_NAME}_sources
   "src/main/java/org/ros2/rcljava/events/EventStatus.java"
   "src/main/java/org/ros2/rcljava/events/PublisherEventStatus.java"
   "src/main/java/org/ros2/rcljava/events/SubscriptionEventStatus.java"
-  "src/main/java/org/ros2/rcljava/events/publisher_statuses/LivelinessLost.java"
-  "src/main/java/org/ros2/rcljava/events/publisher_statuses/OfferedQosIncompatible.java"
+  "src/main/java/org/ros2/rcljava/publisher/statuses/LivelinessLost.java"
+  "src/main/java/org/ros2/rcljava/publisher/statuses/OfferedQosIncompatible.java"
   "src/main/java/org/ros2/rcljava/executors/AnyExecutable.java"
   "src/main/java/org/ros2/rcljava/executors/BaseExecutor.java"
   "src/main/java/org/ros2/rcljava/executors/Executor.java"
